refactor MultiReader into its own file
