+++ webapp/graphite/readers/__init__.py
@@ -1,5 +1,6 @@
 # Import some symbols to avoid breaking compatibility.
-from graphite.readers.utils import BaseReader, CarbonLink, merge_with_cache, MultiReader  # noqa # pylint: disable=unused-import
+from graphite.readers.utils import BaseReader, CarbonLink, merge_with_cache  # noqa # pylint: disable=unused-import
+from graphite.readers.multi import MultiReader  # noqa # pylint: disable=unused-import
 from graphite.readers.whisper import WhisperReader, GzippedWhisperReader  # noqa # pylint: disable=unused-import
 from graphite.readers.ceres import CeresReader  # noqa # pylint: disable=unused-import
 from graphite.readers.rrd import RRDReader  # noqa # pylint: disable=unused-import
